+++ packages/adapter-mongoose/lib/adapter-mongoose.js
@@ -8,19 +8,12 @@ const {
   mapKeyNames,
   identity,
   mergeWhereClause,
-} = require('@keystone-alpha/utils');
+  versionGreaterOrEqualTo,
+} = require('@keystonejs/utils');
 
-const {
-  BaseKeystoneAdapter,
-  BaseListAdapter,
-  BaseFieldAdapter,
-} = require('@keystone-alpha/keystone');
-const {
-  queryParser,
-  pipelineBuilder,
-  mutationBuilder,
-} = require('@keystone-alpha/mongo-join-builder');
-const logger = require('@keystone-alpha/logger').logger('mongoose');
+const { BaseKeystoneAdapter, BaseListAdapter, BaseFieldAdapter } = require('@keystonejs/keystone');
+const { queryParser, pipelineBuilder, mutationBuilder } = require('@keystonejs/mongo-join-builder');
+const logger = require('@keystonejs/logger').logger('mongoose');
 
 const slugify = require('@sindresorhus/slugify');
 
@@ -31,6 +24,7 @@ class MongooseAdapter extends BaseKeystoneAdapter {
     super(...arguments);
     this.name = 'mongoose';
     this.mongoose = new mongoose.Mongoose();
+    this.minVer = '4.0.0';
     if (debugMongoose()) {
       this.mongoose.set('debug', true);
     }
@@ -81,8 +75,24 @@ class MongooseAdapter extends BaseKeystoneAdapter {
 
   getDefaultPrimaryKeyConfig() {
     // Required here due to circular refs
-    const { MongoId } = require('@keystone-alpha/fields-mongoid');
+    const { MongoId } = require('@keystonejs/fields-mongoid');
     return MongoId.primaryKeyDefaults[this.name].getConfig();
+  }
+
+  async checkDatabaseVersion() {
+    let info;
+
+    try {
+      info = await new this.mongoose.mongo.Admin(this.mongoose.connection.db).buildInfo();
+    } catch (error) {
+      console.log(`Error reading version from MongoDB: ${error}`);
+    }
+
+    if (!versionGreaterOrEqualTo(info.versionArray, this.minVer)) {
+      throw new Error(
+        `MongoDB version ${info.version} is incompatible. Version ${this.minVer} or later is required.`
+      );
+    }
   }
 }
 
@@ -210,7 +220,7 @@ class MongooseListAdapter extends BaseListAdapter {
         { include: from.fromField }
       );
       if (ids.length) {
-        args = mergeWhereClause(args, { id: { $in: ids[0][from.fromField] } });
+        args = mergeWhereClause(args, { id: { $in: ids[0][from.fromField] || [] } });
       }
     }
     function graphQlQueryToMongoJoinQuery(query) {