+++ packages/fields/src/types/CloudinaryImage/ImageBlock.js
@@ -1,8 +1,8 @@
 import pluralize from 'pluralize';
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystonejs/build-field-types';
 
-import { Block } from '@keystone-alpha/field-content/Block';
-import { imageContainer, caption } from '@keystone-alpha/field-content/blocks';
+import { Block } from '@keystonejs/field-content/Block';
+import { imageContainer, caption } from '@keystonejs/field-content/blocks';
 import CloudinaryImage from './';
 import SelectType from '../Select';
 import RelationshipType from '../Relationship';