+++ packages/fields/src/types/Unsplash/index.js
@@ -1,4 +1,4 @@
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystonejs/build-field-types';
 import { Unsplash, MongoUnsplashInterface, KnexUnsplashInterface } from './Implementation';
 import { UnsplashBlock } from './UnsplashBlock';